get total items in paging data

+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/manageDukan/content/ManageDukanHeader.kt
@@ -69,7 +69,7 @@ fun ManageDukanHeader(
 
 @Composable
 private fun ProductCountRow(
-    productCount: Int,
+    productCount: Long,
     listener: ManageDukanInteractionListener
 ) {
     Row(
